コントローラーの共通化でconcernsを使用&再度rubocopに通した
