+++ app/controllers/comments_controller.rb
@@ -1,45 +1,5 @@
 # frozen_string_literal: true
 
 class CommentsController < ApplicationController
-  # POST /reports/1/comments
-  # POST /books/1/comments
   before_action :set_comment, only: %i[edit update destroy]
-
-  def edit; end
-
-  def create
-    @comment = @commentable.comments.create(comment_params)
-    @comment.user = current_user
-
-    if @comment.save
-      redirect_to @commentable, notice: t('controllers.common.notice_create', name: Comment.model_name.human)
-    else
-      redirect_to @commentable, notice: t('errors.format', attribute: Comment.model_name.human, message: t('errors.messages.empty'))
-    end
-  end
-
-  def update
-    respond_to do |format|
-      if @comment.update(comment_params)
-        format.html { redirect_to @commentable, notice: t('controllers.common.notice_update', name: Comment.model_name.human) }
-      else
-        format.html { render :edit, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  def destroy
-    @comment.destroy
-    redirect_to @commentable, notice: t('controllers.common.notice_destroy', name: Comment.model_name.human)
-  end
-
-  private
-
-  def set_comment
-    @comment = current_user.comments.find(params[:id])
-  end
-
-  def comment_params
-    params.require(:comment).permit(:content)
-  end
 end
