+++ app/controllers/concerns/commentable.rb
@@ -1,0 +1,43 @@
+# frozen_string_literal: true
+
+module Commentable
+  extend ActiveSupport::Concern
+
+  def edit; end
+
+  def create
+    @comment = @commentable.comments.create(comment_params)
+    @comment.user = current_user
+
+    if @comment.save
+      redirect_to @commentable, notice: t('controllers.common.notice_create', name: Comment.model_name.human)
+    else
+      redirect_to @commentable, notice: t('errors.format', attribute: Comment.model_name.human, message: t('errors.messages.empty'))
+    end
+  end
+
+  def update
+    respond_to do |format|
+      if @comment.update(comment_params)
+        format.html { redirect_to @commentable, notice: t('controllers.common.notice_update', name: Comment.model_name.human) }
+      else
+        format.html { render :edit, status: :unprocessable_entity }
+      end
+    end
+  end
+
+  def destroy
+    @comment.destroy
+    redirect_to @commentable, notice: t('controllers.common.notice_destroy', name: Comment.model_name.human)
+  end
+
+  private
+
+  def set_comment
+    @comment = current_user.comments.find(params[:id])
+  end
+
+  def comment_params
+    params.require(:comment).permit(:content)
+  end
+end
